requirements.txt
asgiref==3.5.2
astroid==2.11.6
black==22.6.0
cached-property==1.5.2
certifi==2020.12.5
cffi==1.15.0
cfgv==3.3.1
chardet==4.0.0
click==8.0.1
colorama==0.4.5
contourpy==1.0.7
cryptography==40.0.2
cycler==0.10.0
decorator==5.0.9
defusedxml==0.7.1
diff-match-patch==20230430
dill==0.3.5.1
distlib==0.3.4
Django==4.1
django-admin-kit==0.0.20
# django-auditlog==2.3.0
django-composite-foreignkey-2==3.9.0
django-cors-headers==3.7.0
django-crontab==0.7.1
django-filter==22.1
django-import-export==3.2.0
django-mssql-backend==2.8.1
django-rest-auth==0.9.5
django-tracking2==0.5.1
django4-background-tasks==1.2.7
djangorestframework==3.13.1
djangorestframework-simplejwt==4.7.1
dnspython==2.1.0
drf-api-tracking==1.8.0
et-xmlfile==1.1.0
exchangelib==4.5.0
filelock==3.7.1
fonttools==4.39.3
geographiclib==2.0
geopy==2.3.0
greenlet==1.1.0
identify==2.5.1
idna==2.10
importlib-metadata==4.0.1
isodate==0.6.0
isort==5.10.1
Jinja2==3.0.1
jsonfield==3.1.0
kiwisolver==1.3.1
lazy-object-proxy==1.7.1
lxml==4.6.3
Markdown==3.3.4
MarkupPy==1.14
MarkupSafe==2.0.1
matplotlib==3.7.0
mccabe==0.7.0
mypy-extensions==0.4.3
nodeenv==1.7.0
ntlm-auth==1.5.0
numpy==1.21.6
numpy-financial==1.0.0
oauthlib==3.1.1
odfpy==1.4.1
openpyxl==3.0.7
packaging==23.1
pandas==1.3.5
pathspec==0.9.0
pbr==5.6.0
pdfminer.six==20201018
Pillow==9.5.0
pip-autoremove==0.10.0
platformdirs==2.5.2
plotly==4.8.1
pre-commit==2.19.0
psycopg2-binary==2.9.6
PuLP==2.6.0
pycparser==2.20
Pygments==2.10.0
PyJWT==2.1.0
pylint==2.14.3
pylint-django==2.5.3
pylint-plugin-utils==0.7
pyodbc==4.0.39
pyparsing==2.4.7
PyPDF2==1.26.0
python-dateutil==2.8.2
pytz==2021.1
PyYAML==6.0
requests==2.25.1
requests-ntlm==1.1.0
requests-oauthlib==1.3.0
retrying==1.3.3
scipy==1.10.1
seaborn==0.11.1
seawater==3.3.4
selenium==3.141.0
simplejwt==2.0.1
six==1.15.0
sortedcontainers==2.4.0
SQLAlchemy==2.0.4
sqlalchemy-migrate==0.11.0
sqlparse==0.4.1
tablib==3.5.0
Tempita==0.5.2
toml==0.10.2
tomli==2.0.1
tomlkit==0.11.0
typing==3.7.4.3
typing_extensions==4.2.0
tzdata==2021.1
tzlocal==3.0
urllib3==1.26.6
virtualenv==20.15.0
whitenoise==5.3.0
wincertstore==0.2
wrapt==1.14.1
xlrd==2.0.1
XlsxWriter==3.0.3
xlwt==1.3.0
zipp==3.4.1
